Category init

+++ jerry-category/jerry-category-client/src/main/java/io/luan/jerry/category/service/CategoryService.java
@@ -6,5 +6,6 @@ import io.luan.jerry.category.sdo.CategorySDO;
  * Service interfaces for Category
  */
 public interface CategoryService {
-    CategorySDO getCategory(Long categoryId);
+
+    CategorySDO getCategory(int categoryId);
 }
